--idle mode

+++ master.c
@@ -152,6 +152,9 @@ void master_loop(char **argv, char **environ) {
 
 	uint8_t uwsgi_signal;
 
+	time_t last_request_timecheck = 0;
+	uint64_t last_request_count = 0;
+
 #ifdef UWSGI_UDP
 	struct sockaddr_in udp_client;
 	socklen_t udp_len;
@@ -236,6 +239,7 @@ void master_loop(char **argv, char **environ) {
 
 	if (uwsgi.cheap) {
 		uwsgi_add_sockets_to_queue(uwsgi.master_queue);
+		uwsgi_log("cheap mode enabled: waiting for socket connection...\n");
 	}
 
 #ifdef UWSGI_MULTICAST
@@ -410,6 +414,7 @@ void master_loop(char **argv, char **environ) {
 		if (uwsgi.restore_snapshot) {
 			uwsgi_log("[snapshot] restoring workers...\n");
 			for(i=1;i<=uwsgi.numproc;i++) {
+				if (uwsgi.workers[i].pid == 0) continue;
 				kill(uwsgi.workers[i].pid, SIGKILL);
 				if (waitpid(uwsgi.workers[i].pid, &waitpid_status, 0) < 0) {
                                 	uwsgi_error("waitpid()");
@@ -916,10 +921,33 @@ void master_loop(char **argv, char **environ) {
 
 			if (uwsgi.numproc > 1) {
 				tmp_counter = 0;
-				for(i=1;i<uwsgi.numproc+1;i++) {
+				for(i=1;i<uwsgi.numproc+1;i++)
 					tmp_counter += uwsgi.workers[i].requests;
-				}
 				uwsgi.workers[0].requests = tmp_counter;
+				if (uwsgi.idle > 0 && !uwsgi.cheap) {
+					uwsgi.current_time = time(NULL);
+					if (!last_request_timecheck) last_request_timecheck = uwsgi.current_time;
+					if (last_request_count != uwsgi.workers[0].requests) {
+						last_request_timecheck = uwsgi.current_time;
+						last_request_count = uwsgi.workers[0].requests;
+					}
+					else if (uwsgi.current_time - last_request_timecheck > uwsgi.idle) {
+						uwsgi_log("workers have been inactive for more than %d seconds\n", uwsgi.idle);
+						for(i=1;i<=uwsgi.numproc;i++) {
+                                			if (uwsgi.workers[i].pid == 0) continue;
+                                			kill(uwsgi.workers[i].pid, SIGKILL);
+                                			if (waitpid(uwsgi.workers[i].pid, &waitpid_status, 0) < 0) {
+                                        			uwsgi_error("waitpid()");
+                                			}
+						}
+						master_has_children = 0;
+						uwsgi.cheap = 1;
+						uwsgi_add_sockets_to_queue(uwsgi.master_queue);
+                				uwsgi_log("cheap mode enabled: waiting for socket connection...\n");
+						last_request_timecheck = 0;
+						continue;	
+					}
+				}
 			}
 
 			// remove expired cache items TODO use rb_tree timeouts
@@ -1171,4 +1199,5 @@ void master_loop(char **argv, char **environ) {
 
 	}
 }
+
 }
